requirements.txt
asgiref==3.7.2
Django==4.2.4
django-nested-admin==4.0.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-yasg==1.21.7
inflection==0.5.1
packaging==23.1
PyJWT==2.8.0
python-monkey-business==1.0.0
pytz==2023.3
PyYAML==6.0.1
six==1.16.0
sqlparse==0.4.4
uritemplate==4.1.1
